requirements.txt
flask
flask-compress
pymongo
sentence-transformers
torch
langchain>=0.3.1
langchain-community
langchain-core
langchain-ollama
langchain-huggingface
faiss-cpu
python-docx
PyPDF2
